S684: WCAG error ARIA (14 files) + #310 discount rules parseFloat fix
- aria-invalid + aria-describedby on all form inputs with inline error states
- Batch A: 9 component files (modals + BrandFollowManager)
- Batch B: 5 page files (auth flow + create-sale)
- #310 color-rules: parseInt -> parseFloat for decimal discount percentages

+++ packages/frontend/pages/organizer/color-rules.tsx
@@ -59,7 +59,7 @@ const ColorRulesPage = () => {
       const payload = {
         tagColor: formData.tagColor,
         label: formData.label,
-        discountPercent: parseInt(formData.discountPercent),
+        discountPercent: parseFloat(formData.discountPercent),
         activeFrom: formData.activeFrom || null,
         activeTo: formData.activeTo || null,
       };
@@ -84,7 +84,7 @@ const ColorRulesPage = () => {
       const payload = {
         tagColor: formData.tagColor,
         label: formData.label,
-        discountPercent: parseInt(formData.discountPercent),
+        discountPercent: parseFloat(formData.discountPercent),
         activeFrom: formData.activeFrom || null,
         activeTo: formData.activeTo || null,
       };
